fix(meta): remove nix from the project (#4)
It's more hassle than it's worth. Reproducable builds are
nice, but the only thing I was using it for was an up
to date record of dependencies. The CI file sort of does that
now anyway.

I know this probably isn't a good practice, but it's fine for
a personal project.